pyproject.toml
[project]
name = "catholic-catechism-rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.120.1",
  "uvicorn==0.38.0",
  "gunicorn==23.0.0",
  "pydantic==2.12.3",
  "qdrant-client==1.15.1",
  "fastembed==0.7.3",
  "python-dotenv==1.2.1",
  "requests==2.32.5",
  "beautifulsoup4==4.14.2",
  "numpy==2.3.4",
  "huggingface-hub>=0.36.2",
  "aiohttp>=3.13.4",
]
